+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.IntegrationTests/IntegrationTests/WebApi/QueryServices/ChargePricesQueryServiceTests.cs
@@ -105,7 +105,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var searchCriteria = new ChargePricesSearchCriteriaV1DtoBuilder()
                 .WithChargeId(charge.Id)
                 .WithFromDateTime(startOfTodayUtc.ToDateTimeOffset())
-                .WithToDateTime(startOfTodayUtc.ToDateTimeOffset().AddDays(3))
+                .WithToDateTime(expectedPoint.Time.ToDateTimeOffset())
                 .WithSkip(2)
                 .Build();
 
@@ -170,7 +170,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var searchCriteria = new ChargePricesSearchCriteriaV1DtoBuilder()
                 .WithChargeId(charge.Id)
                 .WithFromDateTime(InstantHelper.GetTodayAtMidnightUtc().ToDateTimeOffset())
-                .WithToDateTime(InstantHelper.GetTodayPlusDaysAtMidnightUtc(3).ToDateTimeOffset())
+                .WithToDateTime(InstantHelper.GetTodayPlusDaysAtMidnightUtc(2).ToDateTimeOffset())
                 .WithTake(1)
                 .Build();
 
@@ -193,8 +193,8 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var searchCriteria = new ChargePricesSearchCriteriaV1DtoBuilder()
                 .WithChargeId(charge.Id)
                 .WithFromDateTime(charge.Points.First().Time.ToDateTimeOffset())
-                .WithToDateTime(charge.Points.Last().Time.ToDateTimeOffset().AddDays(1))
-                .WithSortColumnName((SortColumnName)(-1))
+                .WithToDateTime(charge.Points.Last().Time.ToDateTimeOffset())
+                .WithSortColumnName((ChargePriceSortColumnName)(-1))
                 .Build();
 
             // Act
@@ -262,8 +262,8 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var searchCriteria = new ChargePricesSearchCriteriaV1DtoBuilder()
                 .WithChargeId(charge.Id)
                 .WithFromDateTime(charge.Points.First().Time.ToDateTimeOffset())
-                .WithToDateTime(charge.Points.Last().Time.ToDateTimeOffset().AddDays(1))
-                .WithSortColumnName(SortColumnName.Price)
+                .WithToDateTime(charge.Points.Last().Time.ToDateTimeOffset())
+                .WithSortColumnName(ChargePriceSortColumnName.Price)
                 .WithIsDescending(false)
                 .Build();
 
@@ -286,8 +286,8 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var searchCriteria = new ChargePricesSearchCriteriaV1DtoBuilder()
                 .WithChargeId(charge.Id)
                 .WithFromDateTime(charge.Points.First().Time.ToDateTimeOffset())
-                .WithToDateTime(charge.Points.Last().Time.ToDateTimeOffset().AddDays(1))
-                .WithSortColumnName(SortColumnName.Price)
+                .WithToDateTime(charge.Points.Last().Time.ToDateTimeOffset())
+                .WithSortColumnName(ChargePriceSortColumnName.Price)
                 .WithIsDescending(true)
                 .Build();
 
@@ -305,8 +305,8 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             // Arrange
             await using var chargesDatabaseWriteContext = _databaseManager.CreateDbContext();
 
-            var expectedPoint = new Point(1.00m, InstantHelper.GetTodayAtMidnightUtc());
-            var unexpectedPoint = new Point(10.00m, InstantHelper.GetTodayPlusDaysAtMidnightUtc(1));
+            var expectedPoint = new Point(1.00m, InstantHelper.GetTomorrowAtMidnightUtc());
+            var unexpectedPoint = new Point(10.00m, InstantHelper.GetTodayPlusDaysAtMidnightUtc(2));
             var points = new List<Point> { expectedPoint, unexpectedPoint };
 
             var expectedCharge = await GetValidCharge(chargesDatabaseWriteContext, points, Resolution.PT15M);
@@ -324,8 +324,8 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
 
             var expected = new ChargePriceV1Dto(
                 expectedPoint.Price,
-                InstantHelper.GetTodayPlusMinutesAtMidnightUtc(0).ToDateTimeOffset(),
-                InstantHelper.GetTodayPlusMinutesAtMidnightUtc(15).ToDateTimeOffset());
+                expectedPoint.Time.ToDateTimeUtc(),
+                expectedPoint.Time.ToDateTimeUtc().AddMinutes(15));
 
             // Act
             var actual = sut.Search(searchCriteria);
@@ -436,7 +436,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var searchCriteria = new ChargePricesSearchCriteriaV1DtoBuilder()
                 .WithChargeId(charge.Id)
                 .WithFromDateTime(todayUtc.ToDateTimeOffset())
-                .WithToDateTime(_zonedDateTimeService.AtStartOfTodayPlusDays(5).ToDateTimeUtc().ToDateTimeOffset())
+                .WithToDateTime(_zonedDateTimeService.AtStartOfTodayPlusDays(4).ToDateTimeUtc().ToDateTimeOffset())
                 .Build();
 
             // Act
@@ -475,7 +475,7 @@ namespace GreenEnergyHub.Charges.IntegrationTests.IntegrationTests.WebApi.QueryS
             var searchCriteria = new ChargePricesSearchCriteriaV1DtoBuilder()
                 .WithChargeId(charge.Id)
                 .WithFromDateTime(startOfMonthUtc.ToDateTimeOffset())
-                .WithToDateTime(_zonedDateTimeService.AtStartOfThisMonthPlusMonths(6).ToDateTimeUtc().ToDateTimeOffset())
+                .WithToDateTime(_zonedDateTimeService.AtStartOfThisMonthPlusMonths(4).ToDateTimeUtc().ToDateTimeOffset())
                 .Build();
 
             // Act